% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{bclCall}
\alias{bclCall}
\title{The bclCall function is a constructor for basecallQC objects.}
\usage{
bclCall(Run, RunMetadata, params = NULL, baseCallMetrics, demultiplexMetrics)
}
\arguments{
\item{Run}{The pun to process}

\item{RunMetadata}{Any run metadata to attach (sata.frame)}
}
\description{
The bclCall function is a constructor for basecallQC objects.
}
