% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processExternalFormats_Functions.R
\name{configParams}
\alias{configParams}
\title{Ini config file reader.}
\usage{
configParams(Config)
}
\arguments{
\item{Config}{Any parameters}
}
\description{
Ini config file reader.
}
