% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIlluminaSamplesheets_Functions.R
\docType{methods}
\name{launchBCLcommands}
\alias{launchBCLcommands}
\title{Illumina Basecalling functions.}
\usage{
launchBCLcommands(bclcommands, launch = TRUE)
}
\arguments{
\item{bclcommands}{BCL comands from readyBasecalling function.}

\item{Run_folders_WithRTA}{Run folders with RTA.complete file.}

\item{config}{Config file (ini format) - see manual for full details.}

\item{Launch}{TRUE or FALSE. TRUE makes QSub call to launch BCL command. TRUE and FALSE return BCL command as character string.}
}
\value{
BCL command as character string.
}
\description{
Launchs Illumina BCL commands using appropriate BioCparallel (eventually).
}
\examples{

library(raster)
library(XML)
subFoldersFull <- "/ifs/data/Hiseq/Runs/161105_D00467_0205_AC9L0AANXX"
Run_folders_WithRTA <- "/ifs/data/Hiseq/Runs/161105_D00467_0205_AC9L0AANXX"
fileLocations <- system.file("extdata",package="basecallQC")

configFile <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
config = data.frame(readIniFile(configFile))
bclcommands <- readyBasecalling(Run_folders_WithRTA,subFoldersFull,config,bclVersion="New")
qsubcommands <- makeQsubs(bclcommands[1],launch=T)
}
\author{
Thomas Carroll and Marian Dore
}
