% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{passFilterBarplot}
\alias{passFilterBarplot}
\alias{demuxBarplot}
\title{Functions to create basemasks for basecalling from Illumina samplesheet.}
\usage{
demuxBarplot(BCLQC, groupBy = c("Lane"))
}
\arguments{
\item{BCLQC}{baseCallQC A  basecall QC object}

\item{groupBy}{Character vector of lane and/or Sample}
}
\value{
ggPlotObject A ggplot2 object.
}
\description{
Parses the Illumina samplesheet from run folders to create standardised
}
\author{
Thomas Carroll and Marian Dore
}
