% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{processConvStats}
\alias{processConvStats}
\title{Converstion stats parser.}
\usage{
processConvStats(ConvStatsXML)
}
\arguments{
\item{ConvStatsXML}{ConversionStats locations.}
}
\value{
A datatable of XML results.
}
\description{
Parses the ConversionStats.xml from Illumina Basecalling
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

convStats <- dir(fileLocations,pattern="ConversionStats.xml",full.names=TRUE)
convProcessed <-processConvStats(convStats)

}
\author{
Thomas Carroll
}
