% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{processDemultiplex}
\alias{processDemultiplex}
\title{Demultiplex parser.}
\usage{
processDemultiplex(demuxStatsXML)
}
\arguments{
\item{demuxStatsXML}{Demultiplex locations.}
}
\value{
A datatable of XML results.
}
\description{
Parses the DemultiplexingStats.xml from Illumina Basecalling
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

demuxStats <- dir(fileLocations,pattern="DemultiplexingStats.xml",full.names=TRUE)
demuxProcessed <-processDemultiplex(demuxStats)

}
\author{
Thomas Carroll
}
