% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allMethods.R
\docType{methods}
\name{readlengths}
\alias{readlengths}
\alias{readlengths.bcl2fastqparams}
\alias{readlengths,BCL2FastQparams-method}
\alias{readlengths,BCL2FastQparams-method}
\title{readlengths
Readlengths as defined runParamaeters.xml}
\usage{
\S4method{readlengths}{BCL2FastQparams}(object)

\S4method{readlengths}{BCL2FastQparams}(object = "BCL2FastQparams")
}
\arguments{
\item{object}{A BCL2FastQparams object}
}
\value{
readlengths <- Readlengths as defined runParamaeters.xml
}
\description{
readlengths
Readlengths as defined runParamaeters.xml
}
\examples{
fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),verbose=T)
readlength <- readlengths(bcl2fastqparams)
}
\author{
Thomas Carroll
}
