% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIlluminaSamplesheets_Functions.R
\docType{methods}
\name{readyBasecalling}
\alias{readyBasecalling}
\title{Illumina Basecalling functions.}
\usage{
readyBasecalling(Run_folders_WithRTA, subFoldersFull, config,
  bclVersion = "New")
}
\arguments{
\item{Run_folders_WithRTA}{Run folders with RTA.complete file.}

\item{subFoldersFull}{All run folders.}

\item{config}{Config file (ini format) - see manual for full details.}

\item{bclVersion}{Character vector of "New" or "Old". Refers to versions prior to 1.2.1 (?) as old.}
}
\value{
Shell commands for basecalling.
}
\description{
Parses the Illumina samplesheet from run folders to create standardised
}
\examples{

library(raster)
library(XML)
subFoldersFull <- "/ifs/data/Hiseq/Runs/161105_D00467_0205_AC9L0AANXX"
Run_folders_WithRTA <- "/ifs/data/Hiseq/Runs/161105_D00467_0205_AC9L0AANXX"
fileLocations <- system.file("extdata",package="basecallQC")

configFile <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
config = data.frame(readIniFile(configFile))
#bclcommands <- readyBasecalling(Run_folders_WithRTA,subFoldersFull,config,bclVersion="New")

}
\author{
Thomas Carroll and Marian Dore
}
