% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{runParameters}
\alias{runParameters}
\title{Parses parameters for illumina basecalling.}
\usage{
runParameters(runParameters = NULL)
}
\arguments{
\item{runParameters}{file path to runParameters.xml .}
}
\value{
A datatable of run parameter.
}
\description{
Creates data.frame of run parameters.
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

runParameters <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
runParametersProcessed <- runParameters(runParameters)

}
\author{
Thomas Carroll
}
