% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{sampleDemuxStats}
\alias{sampleDemuxStats}
\title{Generate per sample summary statistics}
\usage{
sampleDemuxStats(demuxProcessed, plot = T)
}
\arguments{
\item{DemuxStats}{Results from a call to processDemultiplex.}
}
\value{
A datatable of summarised per sample results.
}
\description{
Creates per sample summary statistics from demultiplex results
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

demuxStats <- dir(fileLocations,pattern="DemultiplexingStats.xml",full.names=TRUE)
demuxProcessed <- processDemultiplex(demuxStats)
samplesDemuxStats <- sampleDemuxStats(demuxProcessed)

}
\author{
Thomas Carroll
}
