% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{setBCL2FastQparams}
\alias{setBCL2FastQparams}
\alias{setBCL2FastQparams-setBCL2FastQparams}
\title{Set Parameters for BCL2FastQparameters object.}
\usage{
setBCL2FastQparams(runXML = NULL, config = NULL, runDir = NULL,
  outDir = NULL, verbose = TRUE)
}
\arguments{
\item{runXML}{file path to runParameters.xml ,if not specified
looks in run directory.}

\item{config}{file path to config.ini ,if not specified
looks in run directory.}

\item{runDir}{file path to run directory.}

\item{outDir}{file path to run directory.}

\item{verbose}{TRUE or FALSE. Messages on or off. Warnings/errors persist}
}
\value{
A BCL2FastQparams object.
}
\description{
Parameter class and accessors
}
\examples{
fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
setBCL2FastQparams(runXML,config,runDir=getwd(),verbose=FALSE)
}
