% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{summariseConvStats}
\alias{summariseConvStats}
\title{Generate per sample summary statistics}
\usage{
summariseConvStats(convStatsProcessed)
}
\arguments{
\item{convStatsProcessed}{Results from a call to processConvStats.}
}
\value{
summarisedDemuxStats A datatable of summarised per sample results.
}
\description{
Creates per sample summary statistics from demultiplex results
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

convStats <- dir(fileLocations,pattern="ConversionStats.xml",full.names=TRUE)
convStatsProcessed <- processConvStats(convStats)
summarisedConvStats <- summariseConvStats(convStatsProcessed)

}
\author{
Thomas Carroll
}
