% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R
\docType{methods}
\name{summariseDemuxStats}
\alias{summariseDemuxStats}
\title{Generate per sample summary statistics}
\usage{
summariseDemuxStats(demuxProcessed)
}
\arguments{
\item{demuxProcessed}{Results from a call to processDemultiplex.}
}
\value{
summarisedDemuxStats A datatable of summarised per sample results.
}
\description{
Creates per sample summary statistics from demultiplex results
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

demuxStats <- dir(fileLocations,pattern="DemultiplexingStats.xml",full.names=TRUE)
demuxProcessed <- processDemultiplex(demuxStats)
samplesDemuxStats <- summariseDemuxStats(demuxProcessed)

}
\author{
Thomas Carroll
}
