% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{basecallQC-class}
\alias{basecallQC-class}
\alias{basecallQC}
\alias{basecallQC-basecallQC}
\alias{basecallQC}
\title{The basecallQC object.}
\usage{
basecallQC(bcl2fastqparams, RunMetaData = NULL, sampleSheet = NULL,
  baseCallMetrics = NULL, demultiplexMetrics = NULL)
}
\arguments{
\item{sampleSheet}{fr}

\item{baseCallMetrics}{directory of illumina runfolder.}

\item{demultiplexMetrics}{lr.}

\item{BCL2FastQparams}{Parameters for BCL2FastQC program
and current analysis Run.}

\item{RunMetadata}{Dataframe containing Sample_ID column and user-defined metadata.}

\item{Run}{The pun to process}

\item{RunMetadata}{Any run metadata to attach (sata.frame)}
}
\value{
A basecallQC object.
}
\description{
Objects and methods to handle Illumina BCL inputs and output files.
Provides summary QC statistics for basecalling

The bclCall function is a constructor for basecallQC objects.
}
\examples{

warning("Put example here!")
}
\references{
See \url{http://mrccsc.github.io} for more details on soGGi workflows
}
