% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{basecallQC-class}
\alias{basecallQC-class}
\alias{basecallQC}
\alias{basecallQC-basecallQC}
\title{The basecallQC object.}
\arguments{
\item{runFolder}{directory of illumina runfolder.}
}
\value{
A basecallQC object.
}
\description{
Objects and methods to handle Illumina BCL inputs and output files.
Provides summary QC statistics for basecalling
}
\examples{

warning("Put example here!")
}
\references{
See \url{http://mrccsc.github.io} for more details on soGGi workflows
}
