% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{basecallQC-class}
\alias{basecallQC-class}
\alias{basecallQC}
\alias{basecallQC-basecallQC}
\alias{basecallQC}
\title{The basecallQC object.}
\usage{
basecallQC(bcl2fastqparams, RunMetaData = NULL, sampleSheet = NULL,
  baseCallMetrics = NULL, demultiplexMetrics = NULL)
}
\arguments{
\item{sampleSheet}{fr}

\item{baseCallMetrics}{directory of illumina runfolder.}

\item{demultiplexMetrics}{lr.}

\item{BCL2FastQparams}{Parameters for BCL2FastQC program
and current analysis Run.}

\item{RunMetadata}{Dataframe containing Sample_ID column and user-defined metadata.}

\item{Run}{The pun to process}

\item{RunMetadata}{Any run metadata to attach (sata.frame)}
}
\value{
A basecallQC object.

A basecallQC object.
}
\description{
Objects and methods to handle Illumina BCL inputs and output files.
Provides summary QC statistics for basecalling

The bclCall function is a constructor for basecallQC objects.
}
\examples{

warning("Put example here!")
fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)
}
\references{
See \url{http://mrccsc.github.io} for more details on soGGi workflows
}
