% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{BCL2FastQparams-class}
\alias{BCL2FastQparams-class}
\alias{BCL2FastQparams}
\alias{BCL2FastQparams-BCL2FastQparams}
\title{The Parameters for BCL2FastQparamters object.}
\value{
A BCL2FastQparams object.
}
\description{
Parameter class and accessors for use with basecallQC
}
\references{
See \url{http://mrccsc.github.io} for more details on soGGi workflows
}
