% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{bclCall}
\alias{bclCall}
\alias{basecallQC}
\title{The bclCall function is a constructor for basecallQC objects.}
\usage{
basecallQC(Run = NULL, RunMetadata = NULL, params = NULL,
  sampleSheet = NULL, baseCallMetrics = NULL, demultiplexMetrics = NULL)
}
\arguments{
\item{Run}{The pun to process}

\item{RunMetadata}{Any run metadata to attach (sata.frame)}
}
\description{
The bclCall function is a constructor for basecallQC objects.
}
