% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIlluminaSamplesheets_Functions.R
\docType{methods}
\name{createBasemasks}
\alias{createBasemasks}
\title{Functions to create basemasks for basecalling from Illumina samplesheet.}
\usage{
createBasemasks(cleanedSampleSheet, param = NULL)
}
\arguments{
\item{param}{mn}

\item{sampleSheet}{mm}
}
\value{
basemasks Basemasks.
}
\description{
Parses the Illumina samplesheet from run folders to create standardised
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
cleanedSampleSheet <- validateBCLSheet(sampleSheet,param=NULL)

}
\author{
Thomas Carroll and Marian Dore
}
