% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{demuxBarplot}
\alias{demuxBarplot}
\title{Barplot of Illumina demultiplexing statistics.}
\usage{
demuxBarplot(BCLQC, groupBy = c("Lane"))
}
\arguments{
\item{BCLQC}{baseCallQC A  basecall QC object}

\item{groupBy}{Character vector of lane and/or Sample}
}
\value{
ggPlotObject A ggplot2 object.
}
\description{
Produces a barplot of for demultiplexing statistics of reads with perfect/mismatched barcode.
}
\examples{
fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)
plot <- demuxBarplot(bclQC)
}
\author{
Thomas Carroll and Marian Dore
}
