% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{passFilterBar}
\alias{passFilterBar}
\title{Barplot of Illumina basecalling statistics for reads passing filter.}
\usage{
passFilterBar(BCLQC, groupBy = c("Lane"), metricToPlot = "Yield")
}
\arguments{
\item{BCLQC}{baseCallQC A  basecall QC object}

\item{groupBy}{Character vector defining how plot will be grouped}

\item{metricToPlot}{Character vector defining which metric will be displayed in plot.}
}
\value{
ggPlotObject A ggplot2 object.
}
\description{
Produces a barplot of basecalling statistics for reads passing filter.
}
\examples{
fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)
plot <- passFilterBar(bclQC)
}
\author{
Thomas Carroll and Marian Dore
}
