% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{methods}
\name{setBCL2FastQparams}
\alias{setBCL2FastQparams}
\alias{setBCL2FastQparams-setBCL2FastQparams}
\title{Set Parameters for BCL2FastQparamters object.}
\usage{
setBCL2FastQparams(runXML)
}
\arguments{
\item{runXML}{file path to runParameters.xml}
}
\value{
A BCL2FastQparams object.
}
\description{
Parameter class and accessors
}
\examples{

warning("Put example here!")
}
\references{
See \url{http://mrccsc.github.io} for more details on soGGi workflows
}
