% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processXMLs_Functions.R, R/tables.R
\docType{methods}
\name{summariseConvStats}
\alias{summariseConvStats}
\alias{summariseConvStats}
\alias{summaryDemuxTable}
\title{Generate per sample summary statistics}
\usage{
summariseConvStats(convStatsProcessed, plot = T)

summaryDemuxTable(BCLQC, output = "static")
}
\arguments{
\item{ConvStats}{Results from a call to processConvStats.}

\item{ConvStats}{Results from a call to processConvStats.}
}
\value{
A datatable of summarised per sample results.

A datatable of summarised per sample results.
}
\description{
Creates per sample summary statistics from demultiplex results

Creates per sample summary statistics from demultiplex results
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

convStats <- dir(fileLocations,pattern="ConversionStats.xml",full.names=TRUE)
convStatsProcessed <- processConvStats(convStats)
summarisedConvStats <- summariseConvStats(convStatsProcessed)


fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)
summaryDemuxTable(bclQC,output="static")
}
\author{
Thomas Carroll

Thomas Carroll
}
