% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R, R/tables.R
\docType{methods}
\name{summaryDemuxTable}
\alias{summaryDemuxTable}
\alias{reportBCL}
\alias{summaryDemuxTable}
\title{Generate per sample summary statistics}
\usage{
reportBCL(BCLQC, reportOut = file.path("/Users/tcarroll/genomicsProcessing/",
  "report.html"), output = "static", reportRMDfile = NULL)

summaryDemuxTable(BCLQC, output = "static")
}
\arguments{
\item{ConvStats}{Results from a call to processConvStats.}

\item{ConvStats}{Results from a call to processConvStats.}
}
\value{
A datatable of summarised per sample results.

A datatable of summarised per sample results.
}
\description{
Creates per sample summary statistics from demultiplex results

Creates per sample summary statistics from demultiplex results
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)

fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
outDir <- file.path(fileLocations,"Runs/161105_D00467_0205_AC9L0AANXX/C9L0AANXX/")
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),outDir,verbose=FALSE)
bclQC <- basecallQC(bcl2fastqparams,RunMetaData=NULL,sampleSheet)
summaryDemuxTable(bclQC,output="static")
}
\author{
Thomas Carroll

Thomas Carroll
}
