% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIlluminaSamplesheets_Functions.R
\docType{methods}
\name{validateBCLSheet}
\alias{validateBCLSheet}
\title{Illumina Basecalling functions.}
\usage{
validateBCLSheet(sampleSheet, param = bcl2fastqparams)
}
\arguments{
\item{sampleSheet}{mm}

\item{param}{mn}
}
\value{
Cleaned samplesheet.
}
\description{
Parses the Illumina samplesheet from run folders to create standardised
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")
runXML <- dir(fileLocations,pattern="runParameters.xml",full.names=TRUE)
config <- dir(fileLocations,pattern="config.ini",full.names=TRUE)
sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
bcl2fastqparams <- setBCL2FastQparams(runXML,config,runDir=getwd(),verbose=T)

cleanedSampleSheet <- validateBCLSheet(sampleSheet,param=bcl2fastqparams)

}
\author{
Thomas Carroll and Marian Dore
}
