% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIlluminaSamplesheets_Functions.R
\docType{methods}
\name{validateBCLSheet}
\alias{validateBCLSheet}
\title{Illumina Basecalling functions.}
\usage{
validateBCLSheet(sampleSheet, param = NULL)
}
\arguments{
\item{sampleSheet}{mm}

\item{param}{mn}
}
\value{
Cleaned samplesheet.
}
\description{
Parses the Illumina samplesheet from run folders to create standardised
}
\examples{

fileLocations <- system.file("extdata",package="basecallQC")

sampleSheet <- dir(fileLocations,pattern="*\\\\.csv",full.names=TRUE)
cleanedSampleSheet <- validateBCLSheet(sampleSheet,param=NULL)

}
\author{
Thomas Carroll and Marian Dore
}
